requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# AI & HuggingFace
requests==2.31.0
httpx==0.27.0
Pillow==10.3.0
transformers==4.41.2
torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu
huggingface-hub==0.23.2

# Document Generation
python-docx==1.1.2
python-pptx==0.6.23
reportlab==4.2.0

# Security & Rate Limiting
slowapi==0.1.9
bleach==6.1.0
cryptography==42.0.8

# Monitoring & Logging
prometheus-fastapi-instrumentator==7.0.0
structlog==24.2.0
psutil==5.9.8

# Utilities
jinja2==3.1.4
python-slugify==8.0.4
tenacity==8.3.0
cachetools==5.3.3
